+++ Sources/KeyboardKit/Callouts/Providers/StandardCalloutActionProvider.swift
@@ -35,8 +35,9 @@ open class StandardCalloutActionProvider: CalloutActionProvider {
     }
     
     /**
-     Get the standard action provider, which is used when no
-     custom providers are provided.
+     Get the standard callout action provider.
+
+     This can be set to change the standard value everywhere.
      */
     public static var standardProvider: LocalizedCalloutActionProvider {
         guard let provider = try? EnglishCalloutActionProvider() else { fatalError("EnglishCalloutActionProvider could not be created.") }