+++ Sources/KeyboardKit/Styles/AutocompleteToolbarItemBackgroundStyle.swift
@@ -47,6 +47,8 @@ public extension AutocompleteToolbarItemBackgroundStyle {
     
     /**
      This standard style aims to mimic the native iOS style.
+     
+     This can be set to change the standard value everywhere.
      */
     static var standard = AutocompleteToolbarItemBackgroundStyle()
 }