+++ Sources/KeyboardKit/Styles/AutocompleteToolbarItemStyle.swift
@@ -62,6 +62,8 @@ public extension AutocompleteToolbarItemStyle {
     
     /**
      This standard style aims to mimic the native iOS style.
+
+     This can be set to change the standard value everywhere.
      */
     static var standard = AutocompleteToolbarItemStyle()
 }