+++ Sources/KeyboardKit/Styles/AutocompleteToolbarStyle.swift
@@ -54,6 +54,8 @@ public extension AutocompleteToolbarStyle {
     
     /**
      This standard style aims to mimic the native iOS style.
+
+     This can be set to change the standard value everywhere.
      */
     static var standard = AutocompleteToolbarStyle()
 }