+++ Sources/KeyboardKit/Styles/CalloutStyle.swift
@@ -118,6 +118,8 @@ public extension CalloutStyle {
     
     /**
      This standard style aims to mimic the native iOS style.
+
+     This can be set to change the standard value everywhere.
      */
     static var standard = CalloutStyle()
 }