+++ Sources/KeyboardKit/Styles/InputCalloutStyle.swift
@@ -78,6 +78,8 @@ public extension InputCalloutStyle {
     
     /**
      This standard style aims to mimic the native iOS style.
+
+     This can be set to change the standard value everywhere.
      */
     static var standard = InputCalloutStyle()
 }