+++ Sources/KeyboardKit/Styles/KeyboardButtonBorderStyle.swift
@@ -52,6 +52,8 @@ public extension KeyboardButtonBorderStyle {
     
     /**
      This standard style aims to mimic the native iOS style.
+
+     This can be set to change the standard value everywhere.
      */
     static var standard = KeyboardButtonBorderStyle()
 }