+++ Sources/KeyboardKit/Styles/KeyboardButtonShadowStyle.swift
@@ -54,10 +54,10 @@ public extension KeyboardButtonShadowStyle {
     
     /**
      This standard style aims to mimic the native iOS style.
+
+     This can be set to change the standard value everywhere.
      */
-    static var standard: KeyboardButtonShadowStyle {
-        KeyboardButtonShadowStyle()
-    }
+    static var standard = KeyboardButtonShadowStyle()
 }